fix(GRAFT): _sample_carrier .float() before softmax — bf16 HF logits unsupported by numpy
Mistral 4bit computes in bf16, so organ() returns bf16 logits; numpy has no
bfloat16 dtype and torch.softmax(lg[-1],...).cpu().numpy() raised
"Got unsupported ScalarType BFloat16". The byte-LM organ returns fp32 so this
never surfaced. Fix: lg[-1].float() before softmax. The swap/ablation/
rotation-null bodies already .float() their logits — only this autoregressive
sampling helper missed it. Third (and last) dtype/device gap in the
never-run-on-HF _sample_carrier surfaced by #4527's graft check --hf-model
(instrument-never-run-hides-multiple-bugs).

+++ cli/graft.py
@@ -198,7 +198,7 @@ def _sample_carrier(organ, ids, code, n_bytes, temp=1.0, rng=None):
     with torch.no_grad():
         for _ in range(n_bytes):
             lg = organ(torch.tensor(cur, dtype=torch.long), emb_residual=code)
-            p = torch.softmax(lg[-1], dim=-1).cpu().numpy().astype(np.float64)
+            p = torch.softmax(lg[-1].float(), dim=-1).cpu().numpy().astype(np.float64)
             p = p / p.sum()
             cur.append(int(rng.choice(len(p), p=p)))
     return cur
